app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
google-cloud-secret-manager==2.20.2